Refactor no-gil structs and update test
